+++ docs/pages/components/autocomplete/examples/ExCustomAsync.vue
@@ -3,12 +3,11 @@
         <p class="content"><b>Selected:</b> {{ selected }}</p>
         <b-field label="Find a movie">
             <b-autocomplete
-                v-model="name"
                 :data="data"
                 placeholder="e.g. Fight Club"
                 field="title"
                 :loading="isFetching"
-                @keyup.native="getAsyncData"
+                @typing="getAsyncData"
                 @select="option => selected = option">
 
                 <template slot-scope="props">
@@ -38,7 +37,6 @@
         data() {
             return {
                 data: [],
-                name: '',
                 selected: null,
                 isFetching: false
             }
@@ -46,13 +44,13 @@
         methods: {
             // You have to install and import debounce to use it,
             // it's not mandatory though.
-            getAsyncData: debounce(function () {
-                if (!this.name.length) {
+            getAsyncData: debounce(function (name) {
+                if (!name.length) {
                     this.data = []
                     return
                 }
                 this.isFetching = true
-                this.$http.get(`https://api.themoviedb.org/3/search/movie?api_key=bb6f51bef07465653c3e553d6ab161a8&query=${this.name}`)
+                this.$http.get(`https://api.themoviedb.org/3/search/movie?api_key=bb6f51bef07465653c3e553d6ab161a8&query=${name}`)
                     .then(({ data }) => {
                         this.data = []
                         data.results.forEach((item) => this.data.push(item))