+++ src/components/modal/Modal.vue
@@ -50,7 +50,11 @@
             },
             canCancel: {
                 type: [Array, Boolean],
-                default: () => ['escape', 'x', 'outside', 'button']
+                default: () => {
+                    return config.defaultModalCanCancel
+                        ? config.defaultModalCanCancel
+                        : ['escape', 'x', 'outside', 'button']
+                }
             },
             onCancel: {
                 type: Function,