+++ src/components/taginput/Taginput.vue
@@ -34,7 +34,9 @@
                 :size="size"
                 :disabled="disabled"
                 :loading="loading"
+                :autocomplete="nativeAutocomplete"
                 :keep-first="!allowNew"
+                @typing="onTyping"
                 @focus="onFocus"
                 @blur="customOnBlur"
                 @keydown.native="keydown"
@@ -107,6 +109,7 @@
                 default: 'value'
             },
             autocomplete: Boolean,
+            nativeAutocomplete: String,
             disabled: Boolean,
             ellipsis: Boolean,
             closable: {
@@ -216,10 +219,6 @@
                 this.tags = value
             },
 
-            newTag(value) {
-                this.$emit('typing', value.trim())
-            },
-
             hasInput() {
                 if (!this.hasInput) this.onBlur()
             }
@@ -301,6 +300,10 @@
                     event.preventDefault()
                     this.addTag()
                 }
+            },
+
+            onTyping($event) {
+                this.$emit('typing', $event.trim())
             }
         }
     }