+++ src/components/upload/Upload.vue
@@ -90,8 +90,10 @@
                     this.updateDragDropFocus(false)
                 }
                 const value = event.target.files || event.dataTransfer.files
-                if (value && value.length) {
-                    if (!this.multiple) {
+                if (value) {
+                    if (value.length === 0) {
+                        this.newValue = null
+                    } else if (!this.multiple) {
                         // only one element in case drag drop mode and isn't multiple
                         if (this.dragDrop && value.length !== 1) return false
                         else {